pyproject.toml
[project]
name = "agent1"
version = "0.1.0"
description = "Free, open-source job application automation — powered by Gemma 4 + Playwright"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.24.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "google-genai>=0.1.0",
    "playwright>=1.45.0",
    "httpx>=0.27.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent1"]

[project.scripts]
agent1 = "agent1.cli:app"
